Added inactive redirect to city and post routes

+++ app/controllers/posts_controller.rb
@@ -5,6 +5,7 @@ class PostsController < ApplicationController
   end
 
   def new
+    return if inactive_redirect
     @post = Post.new
     city_id = params[:city_id]
     @city = City.find_by(id: city_id)
@@ -12,6 +13,7 @@ class PostsController < ApplicationController
   end
 
   def create
+    return if inactive_redirect
     city = City.find(params[:city_id])
     new_post = Post.new(post_params)
     user_id = current_user[:id]
@@ -26,6 +28,7 @@ class PostsController < ApplicationController
   end
 
   def show
+    return if inactive_redirect
     post_id = params[:id]
     @post = Post.find_by(id: post_id)
     city_id = params[:city_id]
@@ -33,6 +36,7 @@ class PostsController < ApplicationController
   end
 
   def edit
+    return if inactive_redirect
     post_id = params[:id]
     @post = Post.find_by(id: post_id)
     city_id = params[:city_id]
@@ -41,6 +45,7 @@ class PostsController < ApplicationController
   end
 
   def update
+    return if inactive_redirect
     city = City.find(params[:city_id])
     post_id = params[:id]
     post = Post.find_by(id: post_id)
@@ -56,6 +61,7 @@ class PostsController < ApplicationController
   end
 
   def destroy
+    return if inactive_redirect
     post_id = params[:id]
     post = Post.find_by(id: post_id)
     post.destroy
